pyproject.toml
[project]
name = "langgraph-foundation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.3",
    "python-dotenv>=1.2.2",
]
